[Done] Seção 10: Definindo o Layout e Consumindo o Microsserviço de Produtos

+++ GeekShopping.Web/Services/IServices/IProductService.cs
@@ -5,9 +5,9 @@ namespace GeekShopping.Web.Services.IServices
     public interface IProductService
     {
         Task<IEnumerable<ProductModel>> FindAllProducts();
-        Task<ProductModel> FindProductById(int id);
+        Task<ProductModel> FindProductById(long id);
         Task<ProductModel> CreateProduct(ProductModel model);
         Task<ProductModel> UpdateProduct(ProductModel model);
-        Task<bool> DeleteProductById(int id);
+        Task<bool> DeleteProductById(long id);
     }
 }
